Add a particle eater
It chases particles with the opposite polarity and deletes them when it gets close. When it has deleted a target number, it spawns as many particles of its own polarity and goes dormant for a period.
Currently the result is not very interesting or fun.
In the process, I also generalized movement.