Fix particles disappearing on minimization
This is done by checking requested window dimensions rather than actual, which get put to 0 when minimizing. This will almost certainly have some other effects on the behaviour, but I'll worry about them when I notice them.
Also removed a check from `calculate_force` that was meant to do the same. I don't actually know if that check is unnecessary, but it'll look into it if particles disappear again.